ssh_client: moved to naclsdk v49, reverted glibc-compat changes, cleaned up webports conversion naming
BUG=chromium:687697

Reviewed-on: https://chromium-review.googlesource.com/439207
Tested-by: Mike Frysinger <vapier@chromium.org>

+++ ssh_client/src/syscalls.cc
@@ -250,7 +250,7 @@ int fcntl(int fd, int cmd, ...) {
   return ret;
 }
 
-int ioctl(int fd, long unsigned request, ...) {
+int ioctl(int fd, int request, ...) {
   LOG("ioctl: %d %d\n", fd, request);
   va_list ap;
   va_start(ap, request);
